Refactor: Clean connectors architecture
FINAL STRUCTURE:
- lib/gitprovider.js      → GitProvider base class (new!)
- lib/remotes.js         → Git remote factory (getProvider, detectProvider)
- lib/storage.js         → Storage abstraction
- lib/connectors/       → Shared connector plugins
  - github.js, gitlab.js, bitbucket.js, selfhosted.js
  - pinecone.js

REMOVED:
- lib/connector.js     → unused (connectors/ has all now)
- lib/connectors/git.js → merged into gitprovider.js
- Nested storage/providers folders

Two concerns, one plugin folder:
- remotes.js   → Git hosting (used by connectors/)
- storage.js   → Data storage (can use connectors/)
- connectors/  → Shared plugin implementations
