+++ lib/connectors/bitbucket.js
@@ -1,5 +1,4 @@
-const { GitProvider } = require('./git');
-class BitbucketProvider extends GitProvider {
+class BitbucketProvider extends require("../gitprovider").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'bitbucket' });
         this.token = config.token || process.env.BITBUCKET_TOKEN;
