+++ lib/connectors/github.js
@@ -1,5 +1,4 @@
-const { GitProvider } = require('./git');
-class GitHubProvider extends GitProvider {
+class GitHubProvider extends require("../gitprovider").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'github' });
         this.token = config.token || process.env.GITHUB_TOKEN;
