+++ lib/connectors/gitlab.js
@@ -1,5 +1,4 @@
-const { GitProvider } = require('./git');
-class GitLabProvider extends GitProvider {
+class GitLabProvider extends require("../gitprovider").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'gitlab' });
         this.token = config.token || process.env.GITLAB_TOKEN;
