+++ lib/connectors/selfhosted.js
@@ -1,5 +1,4 @@
-const { GitProvider } = require('./git');
-class SelfHostedProvider extends GitProvider {
+class SelfHostedProvider extends require("../gitprovider").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'selfhosted' });
         this.url = config.url || config.remoteUrl || 'origin';
