+++ lib/remotes.js
@@ -1,4 +1,9 @@
-const { GitProvider } = require('./connectors/git');
+const { execSync } = require('child_process');
+const path = require('path');
+const vaf = require('./vaf');
+const network = require('./network');
+const { GitProvider } = require('./gitprovider');
+
 const { GitHubProvider } = require('./connectors/github');
 const { GitLabProvider } = require('./connectors/gitlab');
 const { BitbucketProvider } = require('./connectors/bitbucket');
@@ -6,9 +11,7 @@ const { SelfHostedProvider } = require('./connectors/selfhosted');
 
 function detectProvider() {
     try {
-        const { execSync } = require('child_process');
         const remoteUrl = execSync('git remote get-url origin', { encoding: 'utf8' }).trim();
-
         if (remoteUrl.includes('github.com') || remoteUrl.includes('github')) {
             return 'github';
         } else if (remoteUrl.includes('gitlab.com') || remoteUrl.includes('gitlab')) {
